create Log

+++ src/main/java/vn/edu/hcmuaf/fit/backend/bookingticket_backend/controller/CityController.java
@@ -1,5 +1,7 @@
 package vn.edu.hcmuaf.fit.backend.bookingticket_backend.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -8,9 +10,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import vn.edu.hcmuaf.fit.backend.bookingticket_backend.dto.CityDTO;
+import vn.edu.hcmuaf.fit.backend.bookingticket_backend.dto.LogDTO;
 import vn.edu.hcmuaf.fit.backend.bookingticket_backend.model.City;
 import vn.edu.hcmuaf.fit.backend.bookingticket_backend.model.Seat;
 import vn.edu.hcmuaf.fit.backend.bookingticket_backend.service.CityService;
+import vn.edu.hcmuaf.fit.backend.bookingticket_backend.service.LogService;
+import vn.edu.hcmuaf.fit.backend.bookingticket_backend.utils.JwtTokenUtils;
 
 import java.io.IOException;
 import java.util.HashMap;
@@ -22,6 +27,10 @@ import java.util.Map;
 @CrossOrigin("http://localhost:3000")
 public class CityController {
     private CityService cityService;
+    @Autowired
+    private JwtTokenUtils jwtTokenUtils;
+    @Autowired
+    private LogService logService;
 
     public CityController(CityService cityService) {
         this.cityService = cityService;
@@ -33,9 +42,17 @@ public class CityController {
 
     // Create a new City
     @PostMapping
-    public ResponseEntity<City> saveCity(@RequestPart("city") CityDTO cityDTO, @RequestPart("file") MultipartFile file) {
+    public ResponseEntity<City> createCity(@RequestPart("city") CityDTO cityDTO, @RequestPart("file") MultipartFile file, HttpServletRequest request) {
         try {
-            City city = cityService.saveCity(cityDTO, file);
+            String token = jwtTokenUtils.extractJwtFromRequest(request);
+            if (token == null) {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+
+            int userId = Integer.parseInt(jwtTokenUtils.extractUserId(token));
+            LogDTO logData =  logService.convertToLogDTO(userId, "Tạo thành phố tên: "+ cityDTO.getName(), 1);
+            logService.createLog(logData);
+            City city = cityService.createCity(cityDTO, file);
             return new ResponseEntity<>(city, HttpStatus.CREATED);
         } catch (IOException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
@@ -68,8 +85,17 @@ public class CityController {
     @PutMapping("{id}")
     public ResponseEntity<City> updateCityById(@PathVariable("id") int id,
                                                @RequestPart("city") CityDTO cityDTO,
-                                               @RequestPart(value = "file", required = false) MultipartFile file) {
+                                               @RequestPart(value = "file", required = false) MultipartFile file,
+                                               HttpServletRequest request) {
         try {
+            String token = jwtTokenUtils.extractJwtFromRequest(request);
+            if (token == null) {
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+
+            int userId = Integer.parseInt(jwtTokenUtils.extractUserId(token));
+            LogDTO logData =  logService.convertToLogDTO(userId, "Cập nhật Thành phố Id: "+ id, 2);
+            logService.createLog(logData);
             City updatedCity = cityService.updateCityByID(cityDTO, file, id);
             return new ResponseEntity<>(updatedCity, HttpStatus.OK);
         } catch (IOException e) {
@@ -79,7 +105,15 @@ public class CityController {
 
     // Delete city by id
     @DeleteMapping("{id}")
-    public ResponseEntity<String> deleteCityById(@PathVariable ("id") int id){
+    public ResponseEntity<String> deleteCityById(@PathVariable ("id") int id, HttpServletRequest request){
+        String token = jwtTokenUtils.extractJwtFromRequest(request);
+        if (token == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
+        int userId = Integer.parseInt(jwtTokenUtils.extractUserId(token));
+        LogDTO logData =  logService.convertToLogDTO(userId, "Xóa thành phố Id: "+ id, 2);
+        logService.createLog(logData);
         cityService.deleteCityByID(id);
         return new ResponseEntity<>("City " + id + " is deleted successfully", HttpStatus.OK);
     }
